+++ examples/provider/provider_token.tf
@@ -1,0 +1,4 @@
+provider "sdwan" {
+  api_token = "my-api-token"
+  url       = "https://10.1.1.1"
+}