+++ CHANGELOG.md
@@ -2,12 +2,15 @@
 
 - Fix issue where `sdwan_policy_group` and `sdwan_configuration_group` resources produce unexpected diffs after import
 - Fix `sdwan_transport_routing_bgp_feature` and `sdwan_service_routing_bgp_feature` resources to correctly handle BGP ASDOT notation for `local_as`, `remote_as` and `as_number` device variables in configuration groups
+- Fix issue where `sdwan_configuration_group` and `sdwan_policy_group` resources trigger a configuration repush to all associated devices when only a single device variable is changed
+- Bump `sdwan_service_dhcp_server_feature` schema to `20.18.0`
 
 ## 0.11.1
 
 - Fix issue where `sdwan_transport_cellular_profile_feature`, `sdwan_system_aaa_feature`, `sdwan_system_ntp_feature`, `sdwan_system_security_feature` and `sdwan_system_snmp_feature` resource produces unexpected changes for the encrypted attributes after the import.
 - Fix issue where `per_tunnel_qos_mode` attribute set to spoke in `sdwan_transport_wan_vpn_interface_ethernet_feature` resource fails.
 - Fix issue where `sdwan_service_lan_vpn_interface_ethernet_feature` fails to be configured if `port_channel_member_interface` is set to `true`
+- Add support for API token authentication
 
 ## 0.11.0